refactor(repositories): extract course-content subquery builders
Split the 7 low-level subquery builders (latest_result, submission_count,
results_count, latest_submission_grade_status, and the message_unread* helpers)
out of course_content_queries.py into a focused course_content_subqueries.py.

These have ZERO external importers -- they are building blocks composed only by
the larger query builders in the same module, which now import them back. So the
move is internally contained: no external import site changes.

course_content_queries.py drops 1275 -> 975 lines; the composite content queries
and per-member count aggregates remain there. Functions moved verbatim (no logic
change); verified both modules compile and import and all 7 names resolve.
